+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - return empty dataframe if `filter` in `get_children` filters out everything (instead 
   of crashing)
 - `update_flexilims` correctly uploads tuples parameters
+- `update_flexilims` correctly uploads floats and np.float/np.int parameters
 
 ### Minor